# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScopeGuard API specification - change order approval workflow for agencies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects for current org
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, COMPLETED, ARCHIVED]
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project with change requests
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project with change requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Archive a project (soft delete)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Archived
  /projects/{id}/export:
    get:
      operationId: exportProjectCsv
      tags: [projects]
      summary: Export approved change requests as CSV
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string

  # CHANGE REQUESTS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]

    # AUTH SCHEMAS
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [ACTIVE, COMPLETED, ARCHIVED]
        hourlyRateCents:
          type: integer
          description: Hourly rate stored in cents
        organizationId:
          type: string
        clientContactId:
          type: string
        clientName:
          type: string
        clientEmail:
          type: string
        clientCompany:
          type: string
        totalApprovedCents:
          type: integer
          description: Sum of all approved change request totals in cents
        changeRequestCount:
          type: integer
        pendingCount:
          type: integer
        approvedCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, description, status, hourlyRateCents, organizationId, clientContactId, clientName, clientEmail, clientCompany, totalApprovedCents, changeRequestCount, pendingCount, approvedCount, createdAt, updatedAt]
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            changeRequests:
              type: array
              items:
                $ref: "#/components/schemas/ChangeRequest"
          required: [changeRequests]
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        clientContactId:
          type: string
        hourlyRateCents:
          type: integer
          minimum: 0
      required: [name, clientContactId, hourlyRateCents]
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        clientContactId:
          type: string
        hourlyRateCents:
          type: integer
          minimum: 0
        status:
          type: string
          enum: [ACTIVE, COMPLETED, ARCHIVED]

    # CHANGE REQUEST SCHEMAS
    ChangeRequest:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        projectName:
          type: string
        clientName:
          type: string
        clientEmail:
          type: string
        title:
          type: string
        description:
          type: string
        estimatedHours:
          type: number
        hourlyRateCents:
          type: integer
        totalCents:
          type: integer
        status:
          type: string
          enum: [DRAFT, SENT, APPROVED, REJECTED, REVISED]
        deadline:
          type: string
          format: date
          nullable: true
        approvalToken:
          type: string
        approvedAt:
          type: string
          format: date-time
          nullable: true
        clientComments:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, projectId, projectName, clientName, clientEmail, title, description, estimatedHours, hourlyRateCents, totalCents, status, approvalToken, clientComments, createdAt, updatedAt]
...